ai-research-agent/requirements.txt
openai>=1.30.0
tavily-python>=0.3.0
truststore>=0.9.0
